Init PhysicsWorld, move static stuff to src files
