+++ include/material.h
@@ -12,10 +12,6 @@ typedef struct {
   GLuint texture_id;
 } TextureEntry;
 
-#define MAX_TEXTURES 128
-static TextureEntry loaded_textures[MAX_TEXTURES];
-static int num_loaded_textures = 0;
-
 struct Texture {
   GLuint texture_id;
   char *texture_type; // Assigned while loading textures (diffuse, specular, etc)
